requirements.txt
# Used for development
black
mypy

# Used for production
Django==3.0.7
djangorestframework==3.10.3
psycopg2